refactor(handlers): split the plugin handler by responsibility
handlers/plugins.py 1002 行同时装着三件不相干的事：通用插件 CRUD、
微信读书专属接口、内置文本工具。通用助手与专属助手混在同一命名空间，
新增插件时「顺手在这里再加一段」阻力最低，文件只会持续变大。

- plugins.py（417 行）只留通用目录、安装、连接、动作与运行历史
- plugin_weread.py（235 行）承接微信读书专属接口
- plugin_booktools.py（476 行）承接三个文本工具的 HTTP 编排
- plugins_common.py 提供共用的请求体解析与错误封装
- 各模块自带 routes()，由 plugins.routes() 聚合，路由与行为完全不变

纯重构，无行为变化；1013 passed 与拆分前一致。

方案：design/webserver/20260824-plugin-capability-interfaces.wip.html （S11 / F-10）
